extract auth and metaStore out of ModelApplication
