+++ src/auth.js
@@ -1,5 +1,8 @@
-define(['facade'], function (facade) {
+define(function (require) {
   'use strict';
+
+  var facade = require('facade');
+  var metaStore = require('store-meta');
 
   var auth = window.BMP.Authentication;
 
@@ -18,4 +21,61 @@ define(['facade'], function (facade) {
       facade.publish('loggedOut');
     });
   });
+
+  auth.getRecord = function (callback) {
+    metaStore.read({
+      id: 'offlineLogin'
+    }).then(
+      function (data) {
+        callback(null, data.attributes);
+      },
+      function (err) {
+        callback(err);
+      }
+    );
+  };
+  auth.setRecord = function (data, callback) {
+    var model = {};
+    model.id = 'offlineLogin';
+    model.toJSON = function () {
+      return model.attributes;
+    };
+    model.attributes = data;
+    model.attributes._id = model.id;
+
+    metaStore.read({
+      id: model.id
+    }).then(
+      function () {
+        metaStore.delete({
+          id: model.id
+        }).then(
+          function () {
+            metaStore.create(model).then(
+              function (document) {
+                callback(null, document);
+              },
+              function (err) {
+                callback(err);
+              }
+            );
+          },
+          function (err) {
+            callback(err);
+          }
+        );
+      },
+      function () {
+        metaStore.create(model).then(
+          function (document) {
+            callback(null, document);
+          },
+          function (err) {
+            callback(err);
+          }
+        );
+      }
+    );
+  };
+
 });
