+++ src/store-meta.js
@@ -1,0 +1,11 @@
+/*globals Backbone:false*/
+define(function (require) {
+  'use strict';
+
+  var NAME = window.BMP.BIC.siteVars.answerSpace.toLowerCase() + '-Meta';
+
+  var Data = require('data');
+
+  return new Data(NAME);
+
+});
